requirements.txt
attrs==23.1.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pre-commit==3.6.0
py==1.11.0
pyparsing==3.1.1
pytest==7.4.4
pytest-repeat==0.9.3
pytest-rerunfailures==13.0
selenium==4.16.0
toml==0.10.2
urllib3==1.26.13
PyYAML==6.0.1
requests==2.28.1
py7zr~=0.20.2
tenacity~=8.2.2
yarl~=1.9.2
